Refactor gameObjects.ts helper
