+++ game-backend/src/game-engine/food-spawner/index.ts
@@ -1,6 +1,6 @@
 import { State } from '../../../../darwin-types/State';
 import spawnFood from './spawnFood';
-import { countUnits, countFood } from '../../helpers/gameObjects';
+import { countUnits, countFood } from '../../helper/gameObjects';
 
 const FOOD_TO_UNIT_RATIO = 2;
