+++ game-backend/src/game-engine/state-selectors/food-selector.ts
@@ -1,10 +1,9 @@
 import { State } from '../../../../darwin-types/State';
 import { Food } from '../../../../darwin-types/game-objects/Food';
 import { GAME_OBJECT_TYPES } from '../../../../darwin-types/game-objects/GameObject';
+import { getGameObjectsPerType } from '../../helper/gameObjects';
 
 const selectFoods = (state: State): Food[] => {
-  return state.objectIds
-    .map(objectId => state.objectMap[objectId])
-    .filter(object => object.type === GAME_OBJECT_TYPES.FOOD);
+  return getGameObjectsPerType(state, GAME_OBJECT_TYPES.FOOD);
 };
 export default selectFoods;
