+++ game-backend/src/helper/gameObjects.ts
@@ -1,10 +1,17 @@
 import { State } from '../../../darwin-types/State';
 import GAME_OBJECT_TYPES from '../constants/gameObjects';
+import { GameObject } from '../../../darwin-types/game-objects/GameObject';
 
-export const countGameObjectsPerType = (state: State, type: string): number =>
+export const getGameObjectsPerType = (
+  state: State,
+  type: string
+): GameObject[] =>
   state.objectIds
     .map(id => state.objectMap[id])
-    .filter(obj => obj.type === type).length;
+    .filter(obj => obj.type === type);
+
+export const countGameObjectsPerType = (state: State, type: string): number =>
+  getGameObjectsPerType(state, type).length;
 
 export const countUnits = (state: State): number =>
   countGameObjectsPerType(state, GAME_OBJECT_TYPES.UNIT);
